Subpage: Improved handling of empty section auto-hiding #2832
Subpage hides empty sections (including in conditions like, if all the items in the section have a start date and this hasn't come up yet).

This change makes sure that it hides them in a consistent manner; in editing mode it doesn't hide, in view mode it does, whoever you are. And it doesn't indicate (falsely) in editing mode that the eye icon is closed.